Pull trait definitions for Agent/Client to SDK
