+++ package.json
@@ -22,7 +22,7 @@
   "type": "module",
   "scripts": {
     "generate:json-schema": "cd rust && cargo run --bin generate --features unstable",
-    "generate": "npm run generate:json-schema && npm run generate:ts-schema && npm run format",
+    "generate": "npm run generate:json-schema && npm run format",
     "format": "prettier --write . && cargo fmt",
     "format:check": "prettier --check . && cargo fmt -- --check",
     "spellcheck": "./scripts/spellcheck.sh",
